+++ fm/config/test.py
@@ -10,5 +10,6 @@ Application configuration when running the Test Suite.
 
 
 CELERY_ALWAYS_EAGER = True
+CELERY_EAGER_PROPAGATES_EXCEPTIONS = True
 
 ECHONEST_API_KEY = 'FOO'