+++ tests/serializers/test_spotify.py
@@ -1,0 +1,13 @@
+# Third Party Libs
+from tests.factories.spotify import TrackFactory
+
+# First Party Libs
+from fm.serializers.spotify import TrackSerializer
+
+
+class TestTrackSerializer(object):
+
+    def should_populate_artists_name(self):
+        track = TrackFactory()
+        serialized = TrackSerializer().serialize(track)
+        assert serialized['artists'][0]['name'], 'artist name is empty'